feat(proveedor): implementa servicios para creación, obtención, actualización y eliminación
He implementado las siguientes mejoras y cambios:

- Servicios: Se añadieron y mejoraron los servicios para la creación, obtención, actualización y eliminación de proveedores.
- Estado: Se ajustó la lógica para manejar correctamente los estados de los encargados y repartidores, mostrando solo los nombres cuando el estado es `false`.
- Respuestas: Se optimizó el formato de las respuestas para facilitar el consumo en el frontend.

Estos cambios mejoran la funcionalidad general de la API, asegurando un manejo más robusto de los datos y una experiencia más intuitiva para los usuarios del sistema.

Authored by: Matias Lagos <matias.lagos2101@alumnos.ubiobio.cl>

+++ frontend/src/hooks/categoria/useDeleteCategoria.jsx
@@ -1,15 +1,23 @@
-// src/hooks/useDeleteCategoria.jsx
 import { deleteCategoria } from '@services/categoria.service';
-import { showErrorAlert, showSuccessAlert } from '@helpers/sweetAlert';
+import { showErrorAlert, showSuccessAlert, deleteDataAlert } from '@helpers/sweetAlert';
 
 const useDeleteCategoria = (fetchCategorias) => {
     const handleDelete = async (id) => {
         try {
-            await deleteCategoria(id); // Elimina la categoría directamente
-            fetchCategorias(); // Actualizamos la lista de categorías después de la eliminación
-            showSuccessAlert('Categoría eliminada con éxito');
+            const result = await deleteDataAlert(); // Mostrar alerta para confirmar la eliminación
+            if (result && result.isConfirmed) {
+                const response = await deleteCategoria(id); // Eliminar la categoría directamente
+                if (response.status >= 400) { // Verificar si hubo un error con la eliminación
+                    return showErrorAlert('Error', response.data?.message || 'Error desconocido');
+                }
+                showSuccessAlert('¡Eliminado!', 'La categoría ha sido eliminada correctamente.');
+                fetchCategorias(); // Actualizamos la lista de categorías después de la eliminación
+            } else {
+                showErrorAlert('Cancelado', 'La operación ha sido cancelada.');
+            }
         } catch (error) {
-            showErrorAlert('Error', 'No se pudo eliminar la categoría');
+            console.error('Error al eliminar la categoría:', error);
+            showErrorAlert('Error de referencia', 'La categoría no puede eliminarse porque está siendo utilizada en otra parte.');
         }
     };
